Move mock_setup_entry to conftest
