+++ script/scaffold/templates/config_flow/tests/test_config_flow.py
@@ -1,5 +1,4 @@
 """Test the NEW_NAME config flow."""
-from collections.abc import Generator
 from unittest.mock import AsyncMock, patch
 
 import pytest
@@ -10,14 +9,7 @@ from homeassistant.components.NEW_DOMAIN.const import DOMAIN
 from homeassistant.core import HomeAssistant
 from homeassistant.data_entry_flow import FlowResultType
 
-
-@pytest.fixture(autouse=True)
-def mock_setup_entry() -> Generator[AsyncMock, None, None]:
-    """Override async_setup_entry."""
-    with patch(
-        "homeassistant.components.NEW_DOMAIN.async_setup_entry", return_value=True
-    ) as mock_setup_entry:
-        yield mock_setup_entry
+pytestmark = pytest.mark.usefixtures("mock_setup_entry")
 
 
 async def test_form(hass: HomeAssistant, mock_setup_entry: AsyncMock) -> None:
