+++ tests/components/onewire/conftest.py
@@ -1,5 +1,6 @@
 """Provide common 1-Wire fixtures."""
-from unittest.mock import MagicMock, patch
+from collections.abc import Generator
+from unittest.mock import AsyncMock, MagicMock, patch
 
 from pyownet.protocol import ConnError
 import pytest
@@ -12,6 +13,15 @@ from homeassistant.core import HomeAssistant
 from .const import MOCK_OWPROXY_DEVICES
 
 from tests.common import MockConfigEntry
+
+
+@pytest.fixture(autouse=True)
+def mock_setup_entry() -> Generator[AsyncMock, None, None]:
+    """Override async_setup_entry."""
+    with patch(
+        "homeassistant.components.onewire.async_setup_entry", return_value=True
+    ) as mock_setup_entry:
+        yield mock_setup_entry
 
 
 @pytest.fixture(name="device_id", params=MOCK_OWPROXY_DEVICES.keys())
