+++ tests/components/philips_js/conftest.py
@@ -1,5 +1,6 @@
 """Standard setup for tests."""
-from unittest.mock import create_autospec, patch
+from collections.abc import Generator
+from unittest.mock import AsyncMock, create_autospec, patch
 
 from haphilipsjs import PhilipsTV
 import pytest
@@ -9,6 +10,17 @@ from homeassistant.components.philips_js.const import DOMAIN
 from . import MOCK_CONFIG, MOCK_ENTITY_ID, MOCK_NAME, MOCK_SERIAL_NO, MOCK_SYSTEM
 
 from tests.common import MockConfigEntry, mock_device_registry
+
+
+@pytest.fixture(autouse=True)
+def mock_setup_entry() -> Generator[AsyncMock, None, None]:
+    """Disable component setup."""
+    with patch(
+        "homeassistant.components.philips_js.async_setup_entry", return_value=True
+    ) as mock_setup_entry, patch(
+        "homeassistant.components.philips_js.async_unload_entry", return_value=True
+    ):
+        yield mock_setup_entry
 
 
 @pytest.fixture(autouse=True)
