+++ tests/components/samsungtv/test_config_flow.py
@@ -1,5 +1,4 @@
 """Tests for Samsung TV config flow."""
-from collections.abc import Generator
 import socket
 from unittest.mock import ANY, AsyncMock, Mock, call, patch
 
@@ -216,14 +215,7 @@ DEVICEINFO_WEBSOCKET_NO_SSL = {
     "timeout": TIMEOUT_WEBSOCKET,
 }
 
-
-@pytest.fixture(autouse=True)
-def mock_setup_entry() -> Generator[AsyncMock, None, None]:
-    """Override async_setup_entry."""
-    with patch(
-        "homeassistant.components.samsungtv.async_setup_entry", return_value=True
-    ) as mock_setup_entry:
-        yield mock_setup_entry
+pytestmark = pytest.mark.usefixtures("mock_setup_entry")
 
 
 @pytest.mark.usefixtures("remote", "rest_api_failing")
