+++ tests/components/sleepiq/test_config_flow.py
@@ -1,5 +1,4 @@
 """Tests for the SleepIQ config flow."""
-from collections.abc import Generator
 from unittest.mock import AsyncMock, patch
 
 from asyncsleepiq import SleepIQLoginException, SleepIQTimeoutException
@@ -12,14 +11,7 @@ from homeassistant.core import HomeAssistant
 
 from .conftest import SLEEPIQ_CONFIG, setup_platform
 
-
-@pytest.fixture(autouse=True)
-def mock_setup_entry() -> Generator[AsyncMock, None, None]:
-    """Override async_setup_entry."""
-    with patch(
-        "homeassistant.components.sleepiq.async_setup_entry", return_value=True
-    ) as mock_setup_entry:
-        yield mock_setup_entry
+pytestmark = pytest.mark.usefixtures("mock_setup_entry")
 
 
 async def test_import(hass: HomeAssistant) -> None:
